Music.Parts.Division

Description

A type to represent (flat) subdivisions of a part.

Documentation

data Division

A division represents a subset of a finite group of performers.

For example a group may be divided into three equal divisions, designated (0, 3), (1, 3) and (2, 3) respectively.

division :: Int -> Int -> Division

Create a division out of a ratio. Dual of getDivision.

getDivision :: Division -> (Int, Int)

divisions :: Int -> [Division]

Get all possible divisions for a given divisor in ascending order.

showDivision :: Division -> String

Show division in ordinary numerals.

showDivisionR :: Division -> String

Show division in roman numerals.